remove syntax taint from ast (as exported via submodule)
Should be no harm in this as the ast is anyway not executed as Racket.
Found suitable code from recursively disarming a syntax tree from
Typed Racket, so using that directly.

This should get rid of the

  "cannot use identifier tainted by macro transformation"

errors that would appear soon as there were macros in
a #lang magnolisp module.