make magnolisp submodules work
this involved switching to resolved-module-path? as the key to
table of modules, to fix mismatch of module path formats used
previously that meant that IDs from submodules would not
resolve to a module in the module table

note the difference between racket apis:
sometimes it's (submod main sub ...+) and sometimes
(main sub ...+); former required for loading, but otherwise
now using resolved-module-path? objects